+++ tests/unit/Parser/KeepAChangeLogTest.php
@@ -10,10 +10,7 @@
 
 namespace ChangeLog\Parser;
 
-use ChangeLog\Log;
-use ChangeLog\Release;
 use Codeception\TestCase\Test;
-use DateTime;
 
 /**
  * Tests for KeepAChangeLog
@@ -133,51 +130,6 @@ class KeepAChangeLogTest extends Test
 		);
 	}
 
-	public function testRender()
-	{
-		$log = new Log;
-		$log->setDescription('This is my change log, it will have lots of changes in it!');
-		$log->setTitle('My change log');
-
-		$release1 = new Release('1.0.0');
-		$release1->setAllChanges([
-			'Added' => ['Thing 1', 'Thing 2'],
-			'Changed' => ['Some change']
-		]);
-		$log->addRelease($release1);
-
-		$release2 = new Release('0.1.0');
-		$release2->setAllChanges([
-			'Added' => ['Initial release'],
-		]);
-		$log->addRelease($release2);
-
-		$expected = <<<'CONTENT'
-# My change log
-This is my change log, it will have lots of changes in it!
-
-## 1.0.0
-### Added
-- Thing 1
-- Thing 2
-
-### Changed
-- Some change
-
-## 0.1.0
-### Added
-- Initial release
-
-CONTENT;
-
-		$actual = $this->parser->render($log);
-		$this->assertEquals(
-			$expected,
-			$actual
-		);
-
-	}
-
 	public function testDate()
 	{
 		$releaseName = '## 1.0.0 - 2015-01-25';
@@ -188,18 +140,6 @@ CONTENT;
 		$this->assertEquals(
 			'2015-01-25',
 			$release->getDate()->format('Y-m-d')
-		);
-	}
-
-	public function testDateRender()
-	{
-		$release = new Release('1.0.0');
-		$release->setDate(DateTime::createFromFormat('Y-m-d', '2015-01-25'));
-
-		$result = $this->parser->renderRelease($release);
-		$this->assertEquals(
-			"\n## 1.0.0 - 2015-01-25",
-			$result
 		);
 	}
 
@@ -234,18 +174,6 @@ CONTENT;
 		);
 	}
 
-	public function testYankedRender()
-	{
-		$release = new Release('1.0.0');
-		$release->setYanked(true);
-
-		$result = $this->parser->renderRelease($release);
-		$this->assertEquals(
-			"\n## 1.0.0 [YANKED]",
-			$result
-		);
-	}
-
 	public function testParseLinks()
 	{
 		$content = [
@@ -267,34 +195,6 @@ CONTENT;
 		$this->assertEquals(
 			'http://fuelphp.com',
 			$log->getRelease('1.0.0')->getLink()
-		);
-	}
-
-	public function testRenderLink()
-	{
-		$log = new Log();
-		$log->setTitle('Change log');
-		$log->setDescription('My change log');
-
-		$release1 = new Release('1.0.0');
-		$release1->setLink('http://fuelphp.com');
-		$log->addRelease($release1);
-
-		$expected = <<<'EXPECTED'
-# Change log
-My change log
-
-## [1.0.0]
-
-[1.0.0] http://fuelphp.com
-
-EXPECTED;
-
-		$result = $this->parser->render($log);
-
-		$this->assertEquals(
-			$expected,
-			$result
 		);
 	}